+++ FarmFresh/FarmFresh.Framework/Entities/Carts/CartItem.cs
@@ -3,13 +3,13 @@ using FarmFresh.Framework.Entities.Products;
 
 namespace FarmFresh.Framework.Entities.Carts
 {
-    public class CartItem : IEntity<int>
+    public class CartItem : IEntity<long>
     {
-        public int ProductId { get; set; }
+        public long ProductId { get; set; }
 
-        public int CartId { get; set; }
+        public long CartId { get; set; }
 
-        public int Quantity { get; set; }
+        public long Quantity { get; set; }
 
         public decimal Cost { get; set; }
