+++ FarmFresh/FarmFresh.Framework/Entities/Carts/CustomerCart.cs
@@ -3,13 +3,13 @@ using FarmFresh.Framework.Entities.Users;
 
 namespace FarmFresh.Framework.Entities.Carts
 {
-    public class CustomerCart : IEntity<int>
+    public class CustomerCart : IEntity<long>
     {
         public Guid CustomerId { get; set; }
 
         public virtual User Customer { get; set; }
 
-        public int CartId { get; set; }
+        public long CartId { get; set; }
 
         public virtual Cart Cart { get; set; }
     }
