+++ FarmFresh/FarmFresh.Framework/Entities/Roles/UserRole.cs
@@ -3,13 +3,13 @@ using FarmFresh.Framework.Entities.Users;
 
 namespace FarmFresh.Framework.Entities.Roles
 {
-    public class UserRole : IEntity<int>
+    public class UserRole : IEntity<long>
     {
         public Guid UserId { get; set; }
 
         public virtual User User { get; set; }
 
-        public int RoleId { get; set; }
+        public long RoleId { get; set; }
 
         public virtual Role Role { get; set; }
     }
