+++ FarmFresh/FarmFresh.Framework/Repositories/Concrete/CustomerCartRepository.cs
@@ -5,7 +5,7 @@ using FarmFresh.Framework.Repositories.Abstract;
 
 namespace FarmFresh.Framework.Repositories.Concrete
 {
-    public class CustomerCartRepository : Repository<CustomerCart, int, FrameworkContext>, ICustomerCartRepository
+    public class CustomerCartRepository : Repository<CustomerCart, long, FrameworkContext>, ICustomerCartRepository
     {
         public CustomerCartRepository(FrameworkContext dbContext)
             : base(dbContext)
