+++ FarmFresh/FarmFresh.Framework/Repositories/Concrete/OrderRepository.cs
@@ -5,7 +5,7 @@ using FarmFresh.Framework.Repositories.Abstract;
 
 namespace FarmFresh.Framework.Repositories.Concrete
 {
-    public class OrderRepository : Repository<Order, int, FrameworkContext>, IOrderRepository
+    public class OrderRepository : Repository<Order, long, FrameworkContext>, IOrderRepository
     {
         public OrderRepository(FrameworkContext dbContext)
             : base(dbContext)
