+++ FarmFresh/FarmFresh.Framework/Repositories/Concrete/ProductCategoryRepository.cs
@@ -5,7 +5,7 @@ using FarmFresh.Framework.Repositories.Abstract;
 
 namespace FarmFresh.Framework.Repositories.Concrete
 {
-    public class ProductCategoryRepository : Repository<ProductCategory, int, FrameworkContext>, IProductCategoryRepository
+    public class ProductCategoryRepository : Repository<ProductCategory, long, FrameworkContext>, IProductCategoryRepository
     {
         public ProductCategoryRepository(FrameworkContext dbContext)
             : base(dbContext)
